Reorganize test folder
